backend/requirements.txt
fastapi==0.110.2
uvicorn==0.29.0
pdfplumber==0.10.3
python-multipart==0.0.9
pydantic==2.11.5
python-dotenv==1.0.1
pdfplumber==0.11.6
openai==1.69.0